+++ src/main/java/app/entities/Orderline.java
@@ -4,14 +4,14 @@ public class Orderline {
 
     private int orderlineId;
     private int orderId;
-    private double initialPrice;
+    private int initialPrice;
 
-    public Orderline(int orderlineId, int orderId, double initialPrice) {
+    public Orderline(int orderlineId, int orderId, int initialPrice) {
         this.orderlineId = orderlineId;
         this.orderId = orderId;
         this.initialPrice = initialPrice;
     }
-    public Orderline(int orderId, double initialPrice) {
+    public Orderline(int orderId, int initialPrice) {
         this.orderId = orderId;
         this.initialPrice = initialPrice;
     }
@@ -32,11 +32,11 @@ public class Orderline {
         this.orderId = orderId;
     }
 
-    public double getInitialPrice() {
+    public int getInitialPrice() {
         return initialPrice;
     }
 
-    public void setInitialPrice(double initialPrice) {
+    public void setInitialPrice(int initialPrice) {
         this.initialPrice = initialPrice;
     }
